requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.6
Django==5.2.12
django-appconf==1.2.0
django-cors-headers==4.9.0
django-cryptography==1.1
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
pycparser==3.0
PyJWT==1.7.1
pytz==2026.1.post1
sqlparse==0.5.5
tzdata==2025.3
